Fraud Detection Engine for E-Commerce/fraudshield/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.2
scikit-learn==1.4.2
python-dotenv==1.0.1
joblib==1.4.0
